Refactor debug build utilities
