+++ src/com/reco1l/BuildConfiguration.kt
@@ -1,0 +1,21 @@
+package com.reco1l
+
+object BuildConfiguration {
+
+    /**
+     * Whether to use textures or not.
+     */
+    const val NO_TEXTURES_MODE = false
+
+    /**
+     * Whether to show entity boundaries or not.
+     */
+    const val SHOW_ENTITY_BOUNDARIES = false
+
+    /**
+     * Whether to use fake multiplayer mode or not.
+     */
+    const val MOCK_MULTIPLAYER = false
+}
+
+
